+++ tests/testthat/test-03-EngineRestGitLab.R
@@ -156,9 +156,9 @@ test_that("`prepare_commits_table()` prepares table of commits properly", {
   )
 })
 
-test_that("`add_repos_issues()` adds issues to repos table", {
+test_that("`get_repos_issues()` adds issues to repos table", {
   gl_repos_by_phrase_table <- test_mocker$use("gl_repos_by_phrase_table")
-  gl_repos_by_phrase_table <- test_rest_priv$add_repos_issues(
+  gl_repos_by_phrase_table <- test_rest_priv$get_repos_issues(
     gl_repos_by_phrase_table
   )
   expect_gt(
@@ -174,9 +174,9 @@ test_that("`add_repos_issues()` adds issues to repos table", {
 
 # public methods
 
-test_that("`add_repos_contributors()` adds contributors to repos table", {
+test_that("`get_repos_contributors()` adds contributors to repos table", {
   expect_snapshot(
-    gl_repos_table <- test_rest$add_repos_contributors(
+    gl_repos_table <- test_rest$get_repos_contributors(
       test_mocker$use("gl_repos_table")
     )
   )