+++ NEWS.md
@@ -1,6 +1,8 @@
+
 GitStats 0.1.0.9002
 
 - set new name for `set_connection` function: `set_host` as it is more informative (and shorter) (I: #271),
+- changed name of a function: `add_team_member` to `set_team_member` (I: #271),
 - added setting tokens by default - if a user does have all the PATs set up in environment variables (as e.g. `GITHUB_PAT` or `GITLAB_PAT`), there is no need to pass them as an arugment to `set_connection` (I: #120 PR: #268),
 - added `get_users()` function to pull information on users (I: #199 PR: #238),
 - added possibility of scanning whole internal git platforms if no `orgs` are passed (I: #258),
@@ -24,4 +26,4 @@ This is the first release of GitStats with given features:
 - `setup()` - setting search parameter to org, team or phrase, setting programming language of repositories,
 - `get_repos()` - pulling repositories from GitHub and GitLab API in a standardized table,
 - `get_commits()` - pulling commits from GitHub and GitLab API in a standardized table,
-- `add_team_member()` - adding team members to GitStats object.
+- `set_team_member()` - adding team members to GitStats object.